Use -source/target 1.6 for Java sources in codegen box tests
Otherwise the corresponding test in `JvmTarget6OnJvm6` (module
`:compiler:tests-different-jdk`) will fail.

The check for IS_SOURCE_6_STILL_SUPPORTED is needed in order to still be
able to run tests in the project while only having a single JDK > 11,
and having all of the env vars JDK_16, JDK_17, JDK_18 pointing to it.